Решены задачи блока 27 - Задачи на права доступа пользователей

+++ delete_profile.php
@@ -5,29 +5,47 @@ try{
 	$dbh = db_connect($config);
 	$active = setSessionUser($dbh);
 	if(!$active){
-		header('Location: login.php', TRUE, 303);
-	}else{
-		if(!empty($_POST)){
-			if($_POST['pass1'] != $_POST['pass2']){
-				$danger = "<p class='alert'> Введенные пароли не совпадают!</p>";
+		header('Location: login.php', TRUE, 303); // Если не запущена сессия, то переходим на старницу авторизации
+	}else{	
+			if(empty($_GET['user'])){
+				$user_delete = $_SESSION['login']; // Если не передан параметром логин пользователя для удаления, то удалять будем свой собственный профиль
 			}else{
-				$active_user = $_SESSION['login'];
-				$user = userDataByLogin($active_user, $dbh);
-				$pass = saltPass($active_user, $_POST['pass1'], $user['salt']);
-				if($pass != $user['password']){
-					$danger = "<p class='alert'> Не правильный пароль пользователя!</p>";
-				}else{
-					$sql_delete_user = 'DELETE FROM users WHERE login=:login';
-					$sth_delete_user = $dbh -> prepare($sql_delete_user);
-					$data['login'] = $active_user;
-					$sth_delete_user -> execute($data);
-					session_unset();
-					session_destroy();
-					header('Location: index.php', TRUE, 303);
+				$user_delete = trim(strip_tags($_GET['user'])); // Если передан, то записываем в переменную $user
+			}
+			
+			if(isAccess('10') || $user == $_SESSION['login']){
+				if(!empty($_POST)){
+					if($_POST['pass1'] != $_POST['pass2']){
+						$danger = "<p class='alert'> Введенные пароли не совпадают!</p>";
+					}else{
+						$active_user = $_SESSION['login'];
+						$user = userDataByLogin($active_user, $dbh);
+						$pass = saltPass($active_user, $_POST['pass1'], $user['salt']);
+						if($pass != $user['password']){
+							$danger = "<p class='alert'> Не правильный пароль пользователя!</p>";
+						}else{
+								
+							$sql_delete_user = 'DELETE FROM users WHERE login=:login';
+							$sth_delete_user = $dbh -> prepare($sql_delete_user);
+							$data_delete_user['login'] = $user_delete;
+							$sth_delete_user -> execute($data_delete_user);
+							if($user_delete == $_SESSION['login']){
+								session_unset();
+								session_destroy();
+								header('Location: index.php', TRUE, 303);
+							}elseif(isAccess('10')){
+								header('Location: admin.php', TRUE, 303);
+							}
+						}
+					}
+				
 				}
+			}else{ // Если страница не твоя или ты не администратор, то выводим предупреждение
+				$danger = "<p class='alert'> У Вас не достаточно прав для удаления пользователя!</p><p class='button'><a href='logout.php'>Выйти</a></p><p class='button'><a href='index.php'>Главная</a></p>";
+				$flag = TRUE;
 			}
 		}
-	}
+
 	?>
 <!DOCTYPE html>
 <html>
@@ -37,10 +55,11 @@ try{
    <meta charset="utf-8">
  </head>
 	 <body>
-	 <h1>Для удаления своего профиля введите пароль</h1>
-	 <?php echo $danger; ?>
+	 <h1>Для удаления профиля введите пароль</h1>
+	 <?php echo $danger; 
+	 if(!$flag){?>
 	 <div class="form">
-		 	<form action="delete_profile.php" method="post">
+		 	<form action="delete_profile.php?user=<?php echo $user_delete; ?>" method="post">
 		 	<p><input type="password" name="pass1">
 		 	<label>Пароль</label></p>
 		 	<p><input type="password"name="pass2">
@@ -48,6 +67,9 @@ try{
 		 	<p><input type="submit" value="Удалить профиль с сайта"></p>
 		 	</form>
 	</div>
+	<?php
+	} 
+	?>
 	 </body>
 </html>
 <?php
